--- src/domain/cafe/gallery/checkGalleryReqService.js
@@ -1,3 +1,0 @@
-export function checkGalleryRequest(req){
-  return req.query.pageNum == null || req.query.category == null || req.query.pageSize == null;
-}
