--- src/domain/cafe/gallery/galleryService.js
@@ -1,8 +1,0 @@
-import { findAllByCategoryInPagination } from '../thumbnail/cafeThumbnailS3.js';
-
-export async function findAllByCategoryForGallery(req){
-  const offset = Number((req.pageNum- 1) * req.pageSize);
-  return await findAllByCategoryInPagination(req.category,req.pageSize,offset);
-}
-
-
