+++ src/domain/cafe/photo/thumbnail/cafeThumbnailS3Service.js
@@ -1,4 +1,4 @@
-import { findAllByCafeUuid as findAllEntitiesByCafeUuid } from "./cafeThumbnailS3.js";
+import { findAllByCafeUuid as findAllEntitiesByCafeUuid, findAllByCategoryInPagination } from "./cafeThumbnailS3.js";
 import {findAll as findAllTumbNails} from "./cafeThumbnailS3.js";
 export async function findAllByCafeUuid(cafeUuid) {
   return await findAllEntitiesByCafeUuid(cafeUuid);
@@ -7,3 +7,10 @@ export async function findAllByCafeUuid(cafeUuid) {
 export async function findAll(){
   return await findAllTumbNails();
 }
+
+export async function findAllByCategoryForGallery(req){
+  const offset = Number((req.pageNum- 1) * req.pageSize);
+  return await findAllByCategoryInPagination(req.category,req.pageSize,offset);
+}
+
+
